ch12/ex05修正

+++ ch12/ex05/index.ts
@@ -8,7 +8,7 @@ export function* readLines(filePath: string) {
   let previouLines = "";
   try {
     while (true) {
-      const bytesRead = readSync(fd, buffer, 0, bufferSize, startPosition);
+      const bytesRead = readSync(fd, buffer as Uint8Array, 0, bufferSize, startPosition);
       if (bytesRead === 0) {
         // 最後の行のyield
         if (previouLines.length > 0) {
